requirements.txt
discord.py
python-dotenv
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client[calendar]==2.108.0
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-core
langgraph
google-generativeai
